refactor: Update navbar structure
