+++ src/App.tsx
@@ -1,16 +1,16 @@
 import './styles/app.scss';
 import './styles/globals.scss';
-import { NavBar } from './components/Organisms/NavBar';
+import { NavBarTemplate } from './components/Templates/NavBarTemplate';
 import { Routes, Route } from 'react-router-dom';
-import { Home } from './components/Templates/Pages/Home';
+import { HomePage } from './components/Templates/Pages/HomePage';
 import { Categories } from './components/Templates/Pages/Categories';
 
 function App() {
   return (
     <>
-      <NavBar />
+      <NavBarTemplate />
       <Routes>
-        <Route path='/' element={<Home />} />
+        <Route path='/' element={<HomePage />} />
         <Route path='/categories/:cat' element={<Categories />} />
       </Routes>
     </>
