+++ src/components/Organisms/NavBar/index.tsx
@@ -1,1 +1,1 @@
-export { default as NavBar } from './NavBar.component';
+export { default as NavBar } from './NavBar'
