+++ src/components/Templates/NavBarTemplate/NavBar.component.tsx
@@ -1,0 +1,12 @@
+import { Logo } from '../../Molucules/Logo';
+import { NavBar } from '../../Organisms/NavBar';
+
+const NavBarTemplate = () => {
+  return (
+    <NavBar>
+      <Logo title='Modern Walk' link='/' />
+    </NavBar>
+  );
+};
+
+export default NavBarTemplate;
