+++ src/components/Templates/NavBarTemplate/index.tsx
@@ -1,0 +1,1 @@
+export { default as NavBarTemplate } from './NavBar.component';
