--- src/components/Templates/Pages/Home/index.tsx
@@ -1,1 +1,0 @@
-export { default as Home } from './Home.component'
