+++ src/components/Templates/Pages/HomePage/index.tsx
@@ -1,0 +1,1 @@
+export { default as HomePage } from './Home.component'
